requirements.txt
pytest==7.1.2
pylint==2.13.9
black==22.3.0
pytest-cov==3.0.0
pydna==4.0.7
pandas==1.4.2 
benchlingapi==2.1.12
numpy==1.22.3
biopython==1.79
synbiopython==0.2.0
python-dotenv==0.20.0
openpyxl==3.0.9